requirements.txt
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
WTForms==2.3.3
Flask-APScheduler==1.12.2
APScheduler==3.7.0
psutil==5.8.0

lxml==4.6.3
beautifulsoup4==4.10.0

requests==2.26.0

Pillow==8.3.2
ImageHash==4.2.1
Distance==0.1.3
ffmpeg-python==0.2.0
filetype==1.0.7

colorama==0.4.4

